fix: component props removing, small changes on request functionality
