+++ packages/pkg.user.settings/components/Secure.tsx
@@ -10,7 +10,6 @@ import { getRelativeTime } from '../utils/getRelativeTime';
 
 export const Secure = () => {
   const user = useMainSt((state) => state.user);
-  const updateUser = useMainSt((state) => state.updateUser);
   const isMobile = useMedia('(max-width: 719px)', false);
 
   const [isChangePasswordModalOpen, setIsChangePasswordModalOpen] = React.useState(false);
@@ -20,12 +19,7 @@ export const Secure = () => {
     <>
       {!isMobile && <h1 className="mb-4 text-3xl font-semibold">Безопасность</h1>}
       <div className="flex flex-col gap-6 sm:gap-8">
-        {!user.emailConfirmed && (
-          <ConfirmEmail
-            allowedConfirmationResend={user.allowedConfirmationResend}
-            updateUser={updateUser}
-          />
-        )}
+        {!user.emailConfirmed && (<ConfirmEmail />)}
         <div className="border-gray-80 flex w-full flex-col rounded-2xl border p-1">
           <div className="flex w-full flex-col p-3">
             <span className="text-xl font-semibold">Данные аккаунта</span>
